move IngestionResourceSnapshot to own file
